pyproject.toml
[project]
name = "personal-portfolio-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "openpyxl",
    "yfinance",
    "python-dotenv",
    "langchain-core",
    "langchain-openai",
    "langgraph",
    "requests>=2.32.5",
    "numpy>=2.4.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]
